fix: build on ARM32 targets (#2063)
* fix(native): Detect libunwind ABI from pointer width

Raspberry Pi OS can run a 32-bit armhf userland with a 64-bit
kernel. In that configuration, CMake reports aarch64 as the system
processor while the compiler emits 32-bit ARM code.

Vendored libunwind previously used the system processor for both the
processor family and ABI width. It consequently combined ARM symbols
with AArch64 DWARF headers and produced this diagnostic:

    vendor/libunwind/src/dwarf/Gexpr.c: In function
    ‘_ULarm_dwarf_eval_expr’:
    vendor/libunwind/include/tdep-aarch64/dwarf-config.h:37:45:
    error: ‘UNW_AARCH64_V31’ undeclared (first use in this function)
    vendor/libunwind/src/dwarf/Gexpr.c:359:35:
    note: in expansion of macro ‘dwarf_to_unw_regnum’

Normalize processor names into x86 and ARM families, then select the
concrete architecture and ELF class from the compiler pointer width.
This also handles the equivalent 32-bit x86 userland configuration.

* fix: Build on baseline ARM32 targets

Avoid aliasing fixed-width module mapping fields as native pointer and
size types in the module finder test. Convert mapped addresses through
uintptr_t before widening them to uint64_t.

Use nop for CPU relaxation on 32-bit ARM, where the baseline assembler
can reject yield as an unsupported instruction. AArch64 continues to
use yield.

Detect whether 64-bit atomic operations require libatomic and link it
when necessary. This supports ARM32 targets where those operations are
not provided directly by the compiler's target architecture.

This fixes the incompatible pointer type, pointer-to-integer cast,
unsupported processor instruction, and unresolved __atomic_load_8 and
__atomic_store_8 diagnostics from the ARM32 build.

* Update CHANGELOG.md

+++ src/modulefinder/sentry_modulefinder_linux.c
@@ -506,7 +506,7 @@ sentry__procmaps_module_to_value(const sentry_module_t *module)
         mmapped_module.is_mmapped = true;
         mmapped_module.num_mappings = 1;
         mmapped_module.mappings[0].addr
-            = (uint64_t)mm.ptr + module->offset_in_inode;
+            = (uint64_t)(uintptr_t)mm.ptr + module->offset_in_inode;
         mmapped_module.mappings[0].size = mm.len - module->offset_in_inode;
 
         sentry__procmaps_read_ids_from_elf(mod_val, &mmapped_module);
